ultralytics/requirements_app.txt
# 柑橘检测Web应用依赖包
streamlit>=1.28.0
plotly>=5.15.0
pandas>=1.5.0
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# 可选依赖（用于更好的性能）
psutil>=5.9.0
GPUtil>=1.4.0
